requirements.txt
modal==1.0.2
azure-storage-blob
python-fasthtml
redis
python-dotenv
loguru
outlines
fastapi
uvicorn
pillow
opencv-python
numpy
pytesseract
transformers
datasets
torch
torchvision
accelerate
requests
pydantic